+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/ServerLevelAccessor.java
@@ -1,0 +1,12 @@
+package insane96mcp.iguanatweaksreborn.mixin;
+
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.level.storage.ServerLevelData;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(ServerLevel.class)
+public interface ServerLevelAccessor {
+    @Accessor
+    ServerLevelData getServerLevelData();
+}