+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/ServerLevelMixin.java
@@ -1,0 +1,15 @@
+package insane96mcp.iguanatweaksreborn.mixin;
+
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import insane96mcp.iguanatweaksreborn.module.sleeprespawn.tiredness.Tiredness;
+import net.minecraft.server.level.ServerLevel;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(ServerLevel.class)
+public abstract class ServerLevelMixin {
+    @ModifyExpressionValue(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/level/ServerLevel;isRaining()Z"))
+    private boolean canSurvive(boolean original) {
+        return Tiredness.onSleepFinished((ServerLevel) (Object) this, original);
+    }
+}